fixed tag drafts not being deleted upon creation / update
